+++ src/legacy/index.ts
@@ -1,6 +1,8 @@
 export * from './Avatar';
 export * from './Button';
 export * from './Checkbox';
+export * from './ModalDialog';
 export * from './Switch';
+export * from './RadioButton';
 export { LegacyTable as LegacyTable } from './Table';
 export * from './Tab';